requirements.txt
pydantic~=2.5.0
Flask~=3.0.0
functions-framework~=3.5.0
google-cloud-secret-manager~=2.17.0
google-cloud~=0.34.0
google-auth~=2.26.1
google-cloud-storage~=2.14.0
firebase_admin~=6.3.0
redivis~=0.14.0